+++ packages/wise_feedback/lib/src/ui/ui.dart
@@ -1,3 +1,4 @@
 export 'feedback_form.dart';
+export 'feedback_toast.dart';
 export 'linear_feedback.dart';
 export 'wise_feedback_theme.dart';
