+++ packages/wise_feedback/test/ui/feedback_toast_test.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:wise_feedback/wise_feedback.dart';
+
+void main() {
+  testWidgets('FeedbackToast renders message and fires onDismiss on tap',
+      (tester) async {
+    var dismissed = 0;
+    await tester.pumpWidget(
+      MediaQuery(
+        data: const MediaQueryData(),
+        child: FeedbackToast(
+          message: 'Saved',
+          isError: false,
+          onDismiss: () => dismissed++,
+        ),
+      ),
+    );
+
+    expect(find.text('Saved'), findsOneWidget);
+    await tester.tap(find.text('Saved'));
+    await tester.pump();
+    expect(dismissed, 1);
+  });
+}
